Revert "* Heeded warning about obsolete Switches in package Prove"
This reverts commit 2415cf8e5f7b8d5fa18dcc02877795e24304eec5.

As it stands right now, used switches for gnatprove are not shown anymore with this option.
Adacore has confirmed that as a bug and it will be fixed in the development version of SPARK,
but until that makes into SPARK community edition (i.e. GNAT/SPARK Community 2020), we just
live with the warning.